+++ crates/mofa-runtime/Cargo.toml
@@ -29,6 +29,7 @@ eyre.workspace = true
 
 # Local dependencies
 mofa-kernel = { path = "../mofa-kernel", version = "0.1", features = ["config"] }
+mofa-foundation = { path = "../mofa-foundation", version = "0.1" }
 mofa-plugins = { path = "../mofa-plugins", version = "0.1" }
 mofa-monitoring = { path = "../mofa-monitoring", version = "0.1", optional = true }
 # Dora-rs dependencies
